feat(breadcrumbs): migrates to material ui
Migrates the Breadcrumbs components to use
material ui

re #143
